fix has column conflict

+++ database/migrations/2026_02_28_025442_add_auth_columns_to_users_table.php
@@ -11,9 +11,12 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('password')->default('');
-            $table->rememberToken();
-            $table->timestamp('email_verified_at')->nullable();
+            if (!Schema::hasColumn('users', 'remember_token')) {
+                $table->rememberToken();
+            }
+            if (!Schema::hasColumn('users', 'email_verified_at')) {
+                $table->timestamp('email_verified_at')->nullable();
+            }
         });
     }
 
@@ -23,9 +26,15 @@ return new class extends Migration {
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('email_verified_at');
-            $table->dropRememberToken();
-            $table->dropColumn('password');
+            if (Schema::hasColumn('users', 'email_verified_at')) {
+                $table->dropColumn('email_verified_at');
+            }
+            if (Schema::hasColumn('users', 'remember_token')) {
+                $table->dropRememberToken();
+            }
+            if (Schema::hasColumn('users', 'password')) {
+                $table->dropColumn('password');
+            }
         });
     }
 };
